scheduler: move k8s framework util out of plugins package tree
The k8s framework/snapshot helper under pkg/scheduler/plugins/util/k8s
is meant to be shared by algorithm plugins and the agent scheduler, but
living under the plugins/ tree made agentscheduler reach into a
plugin-scoped package to use it. Relocate it to
pkg/scheduler/util/k8s so both consumers depend on a common,
non-plugin-specific location instead of agentscheduler depending on
the plugins subtree.
